Accept MODE_DEFAULT for ACCESS_NOTIFICATIONS
In commit e081c964ccd08819e9f5e9c64839d7c833d87e10, the mode for
ACCESS_NOTIFICATIONS was changed from MODE_ALLOWED to MODE_DEFAULT. It
means that when noteOp is called, it may now return MODE_DEFAULT.

Change the check on the return value to also accept MODE_DEFAULT. The
actual permission check is done earlier in the three methods, with a
call to the _enforcePermission method.

Also clean up the @EnforcePermission annotations to clarify their scope.
This is a purely esthetic change.

Bug: 332787465
Test: boot & check Notification History
Test: confirm b/329409825 remains fixed
Change-Id: I75a83cc42ab265ad381254c7ecafeb2088e665cc