Remove unnecessary query in IMMS#switchUserOnHandlerLocked() (2nd try)
This is the 2nd attempt of submitting my previous CL [1], which got
reverted [2] due to a boottime regression (Bug 329414077).
We have already addressed the major contributor of the boot time
regression [3], so this CL should be ready to be submitted again.

This is a follow up CL to my previous CL [4], which started using

  InputMethodSettingsRepository

as a on-memory data repository of InputMethodSettings to avoid
unnecessary queries to PackageManagerService again.

With this CL

  InputMethodManagerService#switchUserOnHandlerLocked()

also start using InputMethodSettingsRepository instead of querying
available IMEs for the new user dynamically.

This CL is supposed to improve the IME user switching latency.

 [1]: I9437455cd9ad907a3587b8087c2aa2620ef4c8fd
      102c8614ab7c397f75ca388f5794c7cab720357d
 [2]: Ib1ae7a2ac556799a7243ba2798bae3910a74d297
      9444e541734571c9e8a5eb3d27dd12bb6c7ddd30
 [3]: I8e51b1ced4dc16cdca7e898885c64793665fafef
      90dd2cc7f03af29e3c0cbe297d8436c192702e98
 [4]: Ic0dd655fbd86b8ccce2b3298b4c70359a468f9ec
      3b88e174bfadd0bc8daa8dd157a4220bacd7646d

Bug: 309837937
Bug: 329163064
Bug: 333591545
Test: presubmit
Change-Id: If1285c532d70496956938751519810ff008b0eda