Replace IMMS#mSettings with IMMS#mCurrentUserId (2nd try)
This is the 2nd attempt of submitting my previous CLs [1][2], which got
reverted due to a boottime regression (Bug 329414077) [3][4]. We have
already addressed the major contributor of the boot time regression [5],
so this CL should be ready to be submitted again.

This CL removes

  InputMethodManagerService#mSettings

as the same instance can be obtained from

  InputMethodSettingsRepository

as long as we keep track of the current user ID.

This is a mechanical refactoring CL. There must be no user observable
behavior change.

 [1]: Idae0fb6d1a9182efe917f81a1d881d488a1b57e5
      054fb8e9822c4b1943197a9176809425d914ff6c
 [2]: I4b7dc1b0e52e9bf25095eb2e3cc520d46b3869e5
      455af2d16485507f46e8552cfa6d573ef7e8d8db
 [3]: I0d65cb9edcdbdaea45b1fa80abfe917b72fb9a50
      9623752ea5761bb6569bb3970d2025ea85b108a7
 [4]: I585e5c6745c40e51f1080b78367175990b2c5884
      3e14070913982ac41afba467756140fba41cba79
 [5]: I8e51b1ced4dc16cdca7e898885c64793665fafef
      90dd2cc7f03af29e3c0cbe297d8436c192702e98

Bug: 305849394
Fix: 329163064
Fix: 333591545
Test: presubmit
Change-Id: Iaa7f08c9081aac9abe0c95d522185632d575d6b8