Use PriorityQueue for tracked Jobs
Previously this code kept an ordered list in a LinkedList.
Using PriorityQueue, with the element ordering as the priority, reduces code
complexity and also reduces runtime complexity (PQ ops are logN).

Randomly noticed this while looking at profiles of l3d cache refills in
system_server. This method was responsible for 0.815% of l3d cache refills.

Change-Id: I0c0bf358e68fef09c8ebcdbe8e5e53ef82cab387