Check if apps are currently in pip before launching split screen
* Which ever app is not pipped will be launched as a single
fullscreen task
* This handles the case where both apps are intent launches (when
nothing is in recents, i.e. app pairs) and task + intent launch (when
non pipped app is in recents). I don't think it's possible to have
task + task launch since the pipped task isn't a selectable split
target
* Add isInPip() check since mTaskInfo in PipOrganizer isn't sufficient
to determine if the app is in pip or not.

Fixes: 323089902
Test: Launched app pair while YT pipped, with second app being in
recents + not in recents. Also launched both apps from overview with
non pipped app being in recents + not recents and also from all apps.
Tested with launching shortcuts (Gmail compose) as well

Change-Id: I930937f331b0078a880fd495a60acb66694d62cb