Fix predictive back animations jump at beginning
We should not call reset() in BackProgressAnimator#onBackStarted, to prevent progress jumps at the beginning of the animation. The reason for the jumps is that inside `reset()` `skipToEnd()` is called. This causes a SpringAnimation-internal flag `mEndRequested` to be flipped to true, which in turn has the effect that on the next frame the spring is immediately set to the requested position. However, by that time, the progress may already have moved a substantial amount, which causes the animation jumps.

Bug: 341257287
Flag: com.android.window.flags.predictive_back_system_anims NEXTFOOD
Test: Manual, i.e. comparing screen recordings and analysing logcat
Change-Id: Ic0b8926d60a0d37a5ead51e130159d7e3219801e